Broken out the whole drag and drop part
